refactor(backend): 重构金融分析模块为模块化架构
将 `backend/app/services/finance_agent.py` 拆分为多个模块化组件：
- `agents/`: NLP解析、报告生成、主编排Agent
- `forecasting/`: 时序分析、Prophet/XGBoost预测器
- `data/`: 数据获取与预处理

更新 `README.md` 文档说明新架构，并调整相关导入路径
