+++ README.md
@@ -16,7 +16,7 @@
 xiaoyi/
 ├── backend/                # 🔧 后端服务
 │   ├── app/
-│   │   ├── api/           # API路由
+│   │   ├── api/             # API路由层
 │   │   │   └── v1/
 │   │   │       ├── endpoints/  # API端点
 │   │   │       │   ├── chat.py       # 对话分析端点
@@ -24,32 +24,42 @@ xiaoyi/
 │   │   │       │   ├── prediction.py # 预测模型 (预留)
 │   │   │       │   └── upload.py     # 文件上传 (预留)
 │   │   │       └── api.py        # API路由汇总
-│   │   ├── core/          # 核心模块
-│   │   │   ├── config.py         # 配置管理
-│   │   │   └── utils.py          # 工具函数
-│   │   ├── models/        # 数据模型
-│   │   │   └── chat.py           # 对话请求模型
-│   │   ├── services/      # 业务逻辑
-│   │   │   └── finance_agent.py  # 金融分析Agent
-│   │   ├── libs/          # 第三方库
-│   │   └── main.py        # 应用入口
-│   ├── requirements.txt   # Python依赖
-│   ├── .env              # 环境变量 (需自行创建)
-│   └── .env.example      # 环境变量模板
+│   │   ├── core/            # 核心模块
+│   │   │   ├── config.py           # 配置管理
+│   │   │   └── utils.py            # 工具函数
+│   │   ├── models/          # Pydantic数据模型
+│   │   │   └── chat.py             # 对话请求模型
+│   │   ├── agents/          # Agent层
+│   │   │   ├── __init__.py
+│   │   │   ├── nlp_agent.py        # NLP解析Agent
+│   │   │   ├── report_agent.py     # 报告生成Agent
+│   │   │   └── finance_agent.py    # 主编排Agent
+│   │   ├── forecasting/     # 预测模型层
+│   │   │   ├── __init__.py
+│   │   │   ├── base.py             # 基础预测器接口
+│   │   │   ├── analyzer.py         # 时序特征分析
+│   │   │   ├── prophet_forecaster.py  # Prophet模型
+│   │   │   └── xgboost_forecaster.py  # XGBoost模型
+│   │   ├── data/            # 数据层
+│   │   │   ├── __init__.py
+│   │   │   └── fetcher.py          # 数据获取与预处理
+│   │   ├── libs/            # 第三方库
+│   │   └── main.py          # 应用入口
+│   ├── requirements.txt     # Python依赖
+│   ├── .env                # 环境变量 (需自行创建)
+│   └── .env.example        # 环境变量模板
 │
-├── frontend/              # 🎨 前端应用
-│   ├── app/              # Next.js页面
-│   ├── components/       # React组件
-│   │   ├── chat/        # 对话组件
-│   │   ├── layout/      # 布局组件
-│   │   └── lib/         # 工具
+├── frontend/                # 🎨 前端应用
+│   ├── app/                # Next.js页面
+│   ├── components/         # React组件
+│   │   ├── chat/          # 对话组件
+│   │   ├── layout/        # 布局组件
+│   │   └── lib/           # 工具
 │   ├── package.json
 │   └── ...
 │
-└── notebooks/            # 📓 实验和开发笔记
-    ├── finance_agent.py  # Agent原型
-    ├── libs/            # 库文件
-    └── *.ipynb          # Jupyter笔记本
+└── notebooks/              # 📓 实验和开发笔记
+    └── *.ipynb            # Jupyter笔记本
 ```
 
 ## 🚀 快速开始
@@ -119,10 +129,35 @@ pnpm dev
 
 ## 🎯 核心功能说明
 
-### 后端模块
+### 后端架构
 
-#### `finance_agent.py` - 金融分析Agent
-完整的金融数据分析管道:
+#### 模块化设计
+后端采用扁平化的模块架构，各层职责清晰：
+
+| 模块 | 路径 | 职责 |
+|------|------|------|
+| **API层** | `app/api/` | HTTP接口路由 |
+| **数据模型** | `app/models/` | 请求/响应Pydantic模型 |
+| **Agent层** | `app/agents/` | AI Agent业务逻辑 |
+| **预测模型** | `app/forecasting/` | 时序预测模型 |
+| **数据层** | `app/data/` | 数据获取与处理 |
+| **核心** | `app/core/` | 配置和工具函数 |
+
+#### Agent 层 (`app/agents/`)
+- `finance_agent.py` - 主编排Agent，协调完整分析流程
+- `nlp_agent.py` - NLP解析，将自然语言转为数据配置
+- `report_agent.py` - 生成专业的分析报告
+
+#### 预测模型层 (`app/forecasting/`)
+- `base.py` - BaseForecaster接口，统一预测器规范
+- `analyzer.py` - 时序特征分析与特征工程
+- `prophet_forecaster.py` - Prophet时序预测模型
+- `xgboost_forecaster.py` - XGBoost机器学习预测
+
+#### 数据层 (`app/data/`)
+- `fetcher.py` - 从AKShare获取金融数据并标准化
+
+#### 分析流程
 1. **NLP解析**: 将自然语言转换为数据配置
 2. **数据获取**: 从AKShare获取股票数据
 3. **时序分析**: 分析趋势、波动性等特征
