+++ backend/app/agents/__init__.py
@@ -1,0 +1,12 @@
+"""
+Agents Module
+=============
+
+AI Agent 层，负责业务逻辑编排
+"""
+
+from .nlp_agent import NLPAgent
+from .report_agent import ReportAgent
+from .finance_agent import FinanceChatAgent
+
+__all__ = ["NLPAgent", "ReportAgent", "FinanceChatAgent"]
