+++ backend/app/data/__init__.py
@@ -1,0 +1,10 @@
+"""
+Data Services
+=============
+
+数据获取与处理模块
+"""
+
+from .fetcher import DataFetcher
+
+__all__ = ["DataFetcher"]
