+++ backend/app/forecasting/__init__.py
@@ -1,0 +1,18 @@
+"""
+Forecasting Module
+==================
+
+时序预测模型层
+"""
+
+from .base import BaseForecaster
+from .analyzer import TimeSeriesAnalyzer
+from .prophet_forecaster import ProphetForecaster
+from .xgboost_forecaster import XGBoostForecaster
+
+__all__ = [
+    "BaseForecaster",
+    "TimeSeriesAnalyzer",
+    "ProphetForecaster",
+    "XGBoostForecaster",
+]
