+++ backend/requirements.txt
@@ -9,3 +9,4 @@ pandas>=2.0.0
 matplotlib>=3.7.0
 prophet
 xgboost
+scikit-learn
